requirements.txt
discord.py
youtube_dl
pynacl
discord-pretty-help